+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
+### 1.9.3 
 
-## 1.9.2
+- Bug fix: https://github.com/p4535992/foundryvtt-pin-cushion/issues/66
+- Bug fix: https://github.com/p4535992/foundryvtt-pin-cushion/issues/52
+
+### 1.9.2
 
 - Removed old setting for "one click creation" because is supported now on v11
 - Add new setting 'Default image icon for on creation of a note' for set a default image when creating a note on the canvas
@@ -8,7 +12,7 @@
 - Add new feature: Auto scale text nameplates of the note (by default is true)
 - Add new feature: Drag not on token layer if you are a GM
 
-## 1.9.0-1
+### 1.9.0-1
 
 - Update to v11